ActivityService 예외처리

+++ services/ActivityService.js
@@ -1,3 +1,5 @@
+const ArgumentError = require("../errors/ArgumentError");
+
 class ActivityService {
   #activityRepository;
 
@@ -6,14 +8,29 @@ class ActivityService {
   }
 
   async createActivity(activity) {
+    //activity id, name, userId
+    if (!activity.id) throw new ArgumentError("activity id is undefined");
+    if (!activity.name) throw new ArgumentError("activity name is undefined");
+    if (!activity.userId)
+      throw new ArgumentError("activity userId is undefined");
+
     return await this.#activityRepository.save(activity);
   }
 
   async deleteActivity(activity) {
+    //activity id, name, userId
+    if (!activity.id) throw new ArgumentError("activity id is undefined");
+    if (!activity.name) throw new ArgumentError("activity name is undefined");
+    if (!activity.userId)
+      throw new ArgumentError("activity userId is undefined");
+
     await this.#activityRepository.remove(activity);
   }
 
   async selectActivities(user) {
+    // user id, email
+    if (!user.id) throw new ArgumentError("user id is undefined");
+    if (!user.email) throw new ArgumentError("user email is undefined");
     return await this.#activityRepository.findAll(user);
   }
 }
